+++ src/core/reference.ts
@@ -1,19 +1,15 @@
-import { Property } from "./property"
-import { addHiddenFinalProp, addHiddenWritableProp } from '../../utils';
-import { MSTAdminisration } from "../../core/mst-node-administration"
 import { isObservableArray, isObservableMap, observable, computed } from "mobx"
-import { resolve, getMSTAdministration, getRelativePathForNodes, isMST, IMSTNode, escapeJsonPath } from "../../core"
-import { invariant, fail } from "../../utils"
-import { getIdentifierAttribute } from "../complex-types/object"
-import { IType, typecheck } from "../type"
+import { getMSTAdministration, getRelativePathForNodes, isMST, IMSTNode } from "./mst-node"
+import { invariant, fail } from "../utils"
+import { getIdentifierAttribute } from "../types/complex-types/object"
+import { IType, typecheck } from "../types/type"
+import { resolve } from "./mst-operations"
 
 export interface IReference {
     $ref: string
 }
 
-// TODO: rename file to reference property
-
-class Reference {
+export class Reference {
     private targetIdAttribute: string
     @observable public identifier: string | null = null
 
@@ -32,7 +28,6 @@ class Reference {
     }
 
     @computed get get() {
-        debugger;
         const { targetIdAttribute, identifier } = this
         if (identifier === null)
             return null
@@ -77,51 +72,5 @@ class Reference {
         if (this.basePath)
             return this.identifier
         return this.identifier ? { $ref: this.identifier } as IReference : null
-    }
-}
-
-export class ReferenceProperty extends Property {
-    constructor(propertyName: string, private type: IType<any, any>, private basePath: string) {
-        super(propertyName)
-    }
-
-    initialize(targetInstance: any, snapshot: any) {
-        const ref = new Reference(targetInstance, this.type, this.basePath, snapshot[this.name])
-        addHiddenFinalProp(targetInstance, this.name + "$value", ref)
-        const self = this
-        Object.defineProperty(targetInstance, this.name, {
-            get: function() {
-                // TODO: factor those functions out to statics
-                getMSTAdministration(this).assertAlive() // Expensive for each read, so optimize away in prod builds!
-                return ref.get
-            },
-            set: function(v) {
-                const node = getMSTAdministration(this)
-                node.assertWritable()
-                const baseValue = ref.identifier
-                ref.setNewValue(v)
-                if (ref.identifier !== baseValue) {
-                    node.emitPatch({
-                        op: "replace",
-                        path: escapeJsonPath(self.name),
-                        value: ref.serialize
-                    }, node)
-                }
-            }
-        })
-        targetInstance[this.name] = snapshot[this.name]
-    }
-
-    serialize(instance: any, snapshot: any) {
-        snapshot[this.name] = (instance[this.name + "$value"] as Reference).serialize()
-    }
-
-    deserialize(instance: any, snapshot: any) {
-        (instance[this.name + "$value"] as Reference).setNewValue(snapshot[this.name])
-    }
-
-    isValidSnapshot(snapshot: any) {
-        // TODO: and check name is string or $ref object
-        return this.name in snapshot
     }
 }
