+++ src/types/complex-types/object.ts
@@ -25,7 +25,7 @@ import { isReferenceFactory } from "../utility-types/reference"
 import { isIdentifierFactory } from "../utility-types/identifier"
 import { Property } from "../property-types/property"
 import { IdentifierProperty } from "../property-types/identifier-property"
-import { ReferenceProperty } from "../property-types/transformed-property"
+import { ReferenceProperty } from "../property-types/reference-property"
 import { ComputedProperty } from "../property-types/computed-property"
 import { ValueProperty } from "../property-types/value-property"
 import { ActionProperty } from "../property-types/action-property"
